Fixed invite flow for email invites
Previously if someone was invited and signed up with Twitter for
example there was no way for the invite to be recognised. Now if they
enter the same email after linking their Twitter as they were invited
with then the invite link will be made and they will have all the
privileges.